Refactor: Centralize state management with React Context
This commit introduces a global state management system using React Context to handle the state of the video grid.

Previously, the `VideoGrid` component was responsible for managing its own state, including the active column index, modal states, and data fetching. This led to prop drilling and made the component complex.

A new `VideoGridContext` has been created to manage this state centrally. The `VideoGrid` and `VerticalFeed` components have been refactored to use this context, which simplifies their implementation and removes the need for passing down numerous props.

This change improves the overall architecture of the application, making it more scalable and easier to maintain.
