+++ components/Providers.tsx
@@ -3,6 +3,7 @@
 import { UserProvider } from '@/context/UserContext';
 import { LanguageProvider } from '@/context/LanguageContext';
 import { ToastProvider } from '@/context/ToastContext';
+import { VideoGridProvider } from '@/context/VideoGridContext';
 import React from 'react';
 
 export function Providers({ children }: { children: React.ReactNode }) {
@@ -10,7 +11,9 @@ export function Providers({ children }: { children: React.ReactNode }) {
     <LanguageProvider>
       <UserProvider>
         <ToastProvider>
-          {children}
+          <VideoGridProvider>
+            {children}
+          </VideoGridProvider>
         </ToastProvider>
       </UserProvider>
     </LanguageProvider>
